llm-test/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
amqp==5.1.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
apispec==6.5.0
appdirs==1.4.4
argcomplete==3.2.3
asgiref==3.7.2
async-timeout==4.0.3
asyncssh==2.14.0
atpublic==4.0
attrs==23.1.0
Babel==2.8.0
backoff==2.2.1
bcrypt==4.1.1
beautifulsoup4==4.12.3
billiard==4.1.0
blinker==1.4
cachelib==0.9.0
cachetools==5.3.2
celery==5.3.4
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.18
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
coloredlogs==15.0.1
colorlog==4.8.0
configobj==5.0.8
configparser==6.0.1
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.2.0
cron-descriptor==1.4.3
croniter==2.0.2
cryptography==42.0.2
cycler==0.12.1
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.6.3
distro==1.7.0
dnspython==2.6.1
docutils==0.20.1
dpath==2.1.6
dulwich==0.21.6
dvc==3.24.0
dvc-data==2.18.1
dvc-http==2.30.2
dvc-objects==1.0.1
dvc-render==0.6.0
dvc-studio-client==0.15.0
dvc-task==0.3.0
effdet==0.4.1
email-validator==1.3.1
emoji==2.10.1
entrypoints==0.4
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.0
fastapi==0.104.1
filelock==3.13.1
filetype==1.2.0
fitz==0.0.1.dev2
flatbuffers==23.5.26
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.49.0
frontend==0.0.3
frozenlist==1.4.0
fsspec==2024.2.0
funcy==2.0
gcovr==5.0
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.25.2
google-re2==1.1
googleapis-common-protos==1.62.0
grandalf==0.8
greenlet==3.0.2
grpcio==1.60.0
gto==1.3.0
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.4
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
hydra-core==1.3.2
idna==3.6
imageio==2.31.5
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflection==0.5.1
iopath==0.1.10
isodate==0.6.1
iterative-telemetry==0.0.8
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==23.5.0
kiwisolver==1.4.5
kombu==5.3.2
kubernetes==28.1.0
langchain-core
langchain-community
langchain
langchainhub==0.1.15
langdetect==1.0.9
launchpadlib==1.10.16
layoutparser==0.3.4
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.10.0
limits==3.10.0
linkify-it-py==2.0.3
llama_cpp_python==0.2.57
lockfile==0.12.2
looseversion==1.3.0
lxml==5.1.0
Mako==1.3.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.7.2
mdit-py-plugins==0.4.0
mdurl==0.1.2
mmh3==4.0.1
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
ordered-set==4.1.0
orjson==3.10.0
overrides==7.4.0
packaging==23.2
pandas==2.2.0
pathlib==1.0.1
pathspec==0.11.2
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pendulum==3.0.0
pikepdf==8.11.0
pillow==10.2.0
pillow_heif==0.15.0
platformdirs==3.11.0
pluggy==1.4.0
portalocker==2.8.2
posthog==3.1.0
prison==0.2.1
progress==1.6
prompt-toolkit==3.0.39
protobuf==4.23.4
prov==2.0.0
psutil==5.9.5
pulsar-client==3.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.26.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydot==1.4.2
pygit2==1.13.1
Pygments==2.16.1
PyGObject==3.42.1
pygtrie==2.5.0
PyJWT==2.3.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pypandoc==1.12
pyparsing==3.0.9
pypdf==4.0.1
pypdfium2==4.27.0
PyPika==0.48.9
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyreadline3==3.4.1
pytesseract==0.3.10
python-daemon==3.0.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
python-nvd3==0.15.0
python-pptx==0.6.23
python-slugify==8.0.4
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rank-bm25==0.2.2
rapidfuzz==3.6.1
razdel==0.5.0
rdflib==7.0.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.6.0
rich-argparse==1.4.0
rpds-py==0.18.0
rsa==4.9
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
safetensors==0.3.2
scikit-learn==1.3.2
scipy==1.10.1
scmrepo==1.3.1
SecretStorage==3.3.1
semver==3.0.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setproctitle==1.3.3
shortuuid==1.0.11
shtab==1.6.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.29
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
sqltrie==0.8.0
ssh-import-id==5.11
starlette==0.27.0
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.2.0
time-machine==2.14.0
timm==0.9.12
tokenizers==0.15.2
tomlkit==0.12.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
traits==6.3.2
transformers==4.37.1
triton==2.2.0
typer==0.9.0
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal-pathlib==0.1.4
unstructured==0.12.6
unstructured-client==0.18.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
vine==5.0.0
voluptuous==0.13.1
wadllib==1.3.6
watchfiles==0.21.0
wcwidth==0.2.8
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==2.2.3
wrapt==1.16.0
WTForms==3.1.2
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.17.0
